Enforce integers in export
Hi,
Thanks for this package. It's most useful.
When using the export capabilities I get random conversions to scientific notation in the output and these are always in the 'start' columns where the `numeric` value 1 has been subtracted from them. This renders the output as unable to be re-imported as the initial scan of the file will spit an error. Using the integer version `1L` should hopefully resolve this.
All the best